refactoring(15): extract useDetailPanelData hook from DetailPanel
Move the three fetch-effect clusters (checkpoint memory, lazy access-log,
lazy node-report + availability probe) and the ancestor-scoped memory
derivation into Tree/useDetailPanelData.ts. The hook is invoked at the
exact same position in the container's hook sequence the inline code
occupied, so effect run-order and fetch/abort timing are unchanged
(§11). DetailPanel 928→794 lines.
