+++ ari-core/ari/viz/frontend/src/README.md
@@ -75,6 +75,7 @@ React/TypeScript source for the ARI `ari.viz` web dashboard — app entry, pages
     - `index.ts` — barrel re-export.
     - `TreePage.tsx` — tree page container.
     - `TreeVisualization.tsx` — D3 search-tree canvas.
+    - `useDetailPanelData.ts` — hook owning DetailPanel's memory/access-log/node-report fetch effects; extracted from DetailPanel in req 15.
     - `DetailPanelTabs/` — extracted detail-panel subcomponents.
       - `README.md` — DetailPanelTabs index.
       - `MemoryEntryCard.tsx` — renders one memory record (own/inherited/global) as a card.
