+++ ari-core/ari/viz/frontend/src/components/Tree/DetailPanel.tsx
@@ -1,16 +1,8 @@
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { useI18n } from '../../i18n';
 import type { TreeNode } from '../../types';
 import { MemoryEntryCard } from './DetailPanelTabs/MemoryEntryCard';
-import { computeAncestorIds } from './detailPanelHelpers';
-import {
-  fetchCheckpointMemory,
-  fetchMemoryAccess,
-  fetchNodeReport,
-  type MemoryEntry,
-  type MemoryAccessResponse,
-  type NodeReport,
-} from '../../services/api';
+import { useDetailPanelData } from './useDetailPanelData';
 
 // ── Colour constants (same as original dashboard.js) ──
 
@@ -49,147 +41,21 @@ export function DetailPanel({ node, allNodes, checkpointId, onClose }: DetailPan
     setActiveTab('overview');
   }, [node?.id]);
 
-  // ── Memory data ──
-  const [memEntries, setMemEntries] = useState<MemoryEntry[] | null>(null);
-  const [globalEntries, setGlobalEntries] = useState<MemoryEntry[]>([]);
-  const [memLoading, setMemLoading] = useState(false);
-  const [memError, setMemError] = useState<string | null>(null);
-
-  // Ancestor chain (root → ... → self) computed from parent_id walk
-  const ancestorIds = useMemo<string[]>(
-    () => computeAncestorIds(node, allNodes),
-    [node, allNodes],
-  );
-
-  useEffect(() => {
-    if (!checkpointId || !node) {
-      setMemEntries(null);
-      return;
-    }
-    let aborted = false;
-    setMemLoading(true);
-    setMemError(null);
-    fetchCheckpointMemory(checkpointId)
-      .then((r) => {
-        if (aborted) return;
-        if (r.error) {
-          setMemError(r.error);
-          setMemEntries([]);
-          setGlobalEntries([]);
-        } else {
-          setMemEntries(r.entries || []);
-          setGlobalEntries(r.global || []);
-        }
-      })
-      .catch((e) => {
-        if (aborted) return;
-        setMemError(String(e));
-        setMemEntries([]);
-      })
-      .finally(() => {
-        if (!aborted) setMemLoading(false);
-      });
-    return () => {
-      aborted = true;
-    };
-  }, [checkpointId, node?.id]);
-
-  const visibleMemory = useMemo(() => {
-    if (!memEntries || !node) return [] as MemoryEntry[];
-    const allowed = new Set(ancestorIds);
-    return memEntries.filter((e) => allowed.has(e.node_id));
-  }, [memEntries, node, ancestorIds]);
-
-  // ── Access log data (lazy: fetched only when the Access tab is opened) ──
-  const [accessData, setAccessData] = useState<MemoryAccessResponse | null>(null);
-  const [accessLoading, setAccessLoading] = useState(false);
-  const [accessError, setAccessError] = useState<string | null>(null);
-
-  useEffect(() => {
-    if (activeTab !== 'access' || !checkpointId || !node) {
-      return;
-    }
-    let aborted = false;
-    setAccessLoading(true);
-    setAccessError(null);
-    fetchMemoryAccess(checkpointId, node.id)
-      .then((r) => {
-        if (aborted) return;
-        if (r.error) setAccessError(r.error);
-        setAccessData(r);
-      })
-      .catch((e) => {
-        if (aborted) return;
-        setAccessError(String(e));
-        setAccessData(null);
-      })
-      .finally(() => {
-        if (!aborted) setAccessLoading(false);
-      });
-    return () => {
-      aborted = true;
-    };
-  }, [activeTab, checkpointId, node?.id]);
-
-  // ── Node report data (lazy: fetched only when the Report tab is opened) ──
-  const [reportData, setReportData] = useState<NodeReport | null>(null);
-  const [reportError, setReportError] = useState<string | null>(null);
-  const [reportLoading, setReportLoading] = useState(false);
-  const [reportAvailable, setReportAvailable] = useState<boolean | null>(null);
-
-  // Probe once per node-id whether a report exists, so we can show/hide the
-  // tab without forcing the user to click it. We hit the same endpoint —
-  // tiny payload — and remember the answer.
-  useEffect(() => {
-    if (!checkpointId || !node) {
-      setReportAvailable(null);
-      setReportData(null);
-      return;
-    }
-    let aborted = false;
-    fetchNodeReport(checkpointId, node.id)
-      .then((r) => {
-        if (aborted) return;
-        if (r.error || !r.report) {
-          setReportAvailable(false);
-          setReportData(null);
-        } else {
-          setReportAvailable(true);
-          setReportData(r.report);
-        }
-      })
-      .catch(() => {
-        if (!aborted) setReportAvailable(false);
-      });
-    return () => {
-      aborted = true;
-    };
-  }, [checkpointId, node?.id]);
-
-  useEffect(() => {
-    if (activeTab !== 'report' || !checkpointId || !node) {
-      return;
-    }
-    if (reportData) return; // already loaded by the availability probe.
-    let aborted = false;
-    setReportLoading(true);
-    setReportError(null);
-    fetchNodeReport(checkpointId, node.id)
-      .then((r) => {
-        if (aborted) return;
-        if (r.error) setReportError(r.error);
-        else if (r.report) setReportData(r.report);
-      })
-      .catch((e) => {
-        if (!aborted) setReportError(String(e));
-      })
-      .finally(() => {
-        if (!aborted) setReportLoading(false);
-      });
-    return () => {
-      aborted = true;
-    };
-  }, [activeTab, checkpointId, node?.id, reportData]);
+  // Per-node data loading (memory / access-log / node-report) lives in a hook.
+  const {
+    ancestorIds,
+    visibleMemory,
+    globalEntries,
+    memLoading,
+    memError,
+    accessData,
+    accessLoading,
+    accessError,
+    reportData,
+    reportError,
+    reportLoading,
+    reportAvailable,
+  } = useDetailPanelData(checkpointId, node, allNodes, activeTab);
 
   // ── Resize drag handlers ──
